+++ .vscode/launch.json
@@ -1,0 +1,18 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "Python: Current File",
+            "type": "python",
+            "request": "launch",
+            "program": "${file}",
+            "args": ["-s", "${workspaceFolder}/supplemental/requests-test.csv"],
+            "cwd": "${workspaceFolder}/batchdoi",
+            "console": "integratedTerminal",
+            "justMyCode": true
+        }
+    ]
+}
