+++ batchdoi/publish.py
@@ -4,47 +4,24 @@
 '''Publish DOIs given input file with DOI names
 '''
 import sys
-import argparse
 import logging
-import json
 import dcdata
 import services
-
 
 import datacite
 
 logger = logging.getLogger(__name__)
 
-def get_args():
-    '''Get command line arguments as well as configuration settings'''
-    parser_desc = 'Publish DOIs.'
-    parser = argparse.ArgumentParser(description=parser_desc)
-    parser.add_argument("doifile", help="Line by line file of DOIs to delete")
-    parser.add_argument("-c", "--config", help="Config file path (defaults to './config.json')", default='config.json')
-    parser.add_argument('-l', '--live', action='store_true', help="Create DOIs on the live system instead of the test system")
-    #parser.add_argument('-v', '--verbose', action='store_true', help="Show detailed output")
-    args = vars(parser.parse_args())
-    return args
 
-def get_config(path):
-    with open(path) as json_data_file:
-        settings = json.load(json_data_file)
-    return settings
-
-def main(argv=None):
-    args = get_args()
-    infile = args['doifile']
-    datacite_params = get_config(args['config'])
-    if args['live']:
-        datacite_settings = datacite_params['datacite_live']
-    else:
-        datacite_settings = datacite_params['datacite_test']
+def main(args, datacite_settings):
     datacite_service = datacite.DataciteService(datacite_settings)
     doi_service = services.DOIService(datacite_service, dcdata.create_publish_payload, None)
+    infile = args['doifile']
     with open(infile) as fh:
         for line in fh:
             doi_name = line.strip()
             doi_service.publish_doi(doi_name)
+
 
 if __name__ == "__main__":
     sys.exit(main())
